Core/Random: Changed random functions returning doubles to return floats
Port From (https://github.com/TrinityCore/TrinityCore/commit/9894f6b802c974bb36acd7fbb0d083455a1f0f1b)

+++ Source/Game/Entities/Unit/Unit.cs
@@ -3393,7 +3393,7 @@ namespace Game.Entities
                     discreteResistProbability[i] = Math.Max(0.5f - 2.5f * Math.Abs(0.1f * i - averageResist), 0.0f);
             }
 
-            float roll = (float)RandomHelper.NextDouble();
+            float roll = RandomHelper.NextSingle();
             float probabilitySum = 0.0f;
 
             uint resistance = 0;
